requirement.txt
apturl==0.5.2
asn1crypto==0.24.0
bcrypt==3.1.7
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docx==0.2.4
fabric==2.5.0
Flask==1.1.1
gpg==1.10.0
httplib2==0.9.2
idna==2.6
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
notify2==0.3
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
pandas==1.0.1
paramiko==2.7.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
PySocks==1.6.5
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-docx==0.8.10
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
torbrowser-launcher==0.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==1.0.0
xkit==0.0.0
zope.interface==4.3.2
